requirements.txt
requests>=2.28
python-dotenv>=1.0
pypdf>=5.0
pypdfium2>=4.30
fastapi>=0.115
uvicorn>=0.30
httpx>=0.27
pytest>=8.0
